Added support for --premajor, --preminor, --prepatch, and --prerelase
