+++ package.json
@@ -1,42 +1,53 @@
 {
- "name": "version-bump-prompt",
- "version": "1.1.0",
- "description": "Automatically (or with prompts) bump your version number, GIT tag, and GIT push",
- "keywords": [
-  "version",
-  "bump",
-  "npm",
-  "node",
-  "bower",
-  "package",
-  "tag",
-  "push",
-  "prompt"
- ],
- "scripts": {
-  "release": "npm-check-updates -u && npm update && bump --prompt --tag --push --all"
- },
- "main": "index.js",
- "bin": {
-  "bump": "./bump.js"
- },
- "license": "MIT",
- "author": "Alexey Raspopov (https://github.com/alexeyraspopov)",
- "homepage": "https://github.com/BigstickCarpet/version-bump-prompt",
- "repository": {
-  "type": "git",
-  "url": "https://github.com/BigstickCarpet/version-bump-prompt.git"
- },
- "dependencies": {
-  "child-process-promise": "^1.0.2",
-  "commander": "^2.8.0",
-  "detect-indent": "^3.0.1",
-  "inquirer": "^0.8.2",
-  "multiline": "^1.0.2",
-  "semver": "^4.3.3"
- },
- "devDependencies": {
-  "npm-check-updates": "^1.5.1",
-  "version-bump-prompt": "^1.0.1"
- }
+  "name": "version-bump-prompt",
+  "version": "1.1.0",
+  "description": "Automatically (or with prompts) bump your version number, GIT tag, and GIT push",
+  "keywords": [
+    "version",
+    "bump",
+    "npm",
+    "node",
+    "bower",
+    "package",
+    "tag",
+    "push",
+    "prompt"
+  ],
+  "scripts": {
+    "semver": "semver",
+    "test": "mocha --recursive tests",
+    "build": "jshint . --verbose && jscs . --verbose",
+    "update-deps": "npm-check-updates -u && npm update",
+    "release": "npm run update-deps && npm test && bump --prompt --tag --push --all && npm publish"
+  },
+  "main": "lib/index.js",
+  "bin": {
+    "bump": "bin/bump.js"
+  },
+  "license": "MIT",
+  "contributors": [
+    "Alexey Raspopov (https://github.com/alexeyraspopov)",
+    "James Messinger (http://JamesMessinger.com)"
+  ],
+  "homepage": "https://github.com/BigstickCarpet/version-bump-prompt",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/BigstickCarpet/version-bump-prompt.git"
+  },
+  "dependencies": {
+    "child-process-promise": "^1.0.2",
+    "commander": "^2.8.1",
+    "detect-indent": "^3.0.1",
+    "inquirer": "^0.8.3",
+    "semver": "^4.3.4"
+  },
+  "devDependencies": {
+    "chai": "^2.3.0",
+    "jscs": "^1.13.1",
+    "jshint": "^2.7.0",
+    "lodash": "^3.8.0",
+    "mocha": "^2.2.4",
+    "npm-check-updates": "^1.5.1",
+    "version-bump-prompt": "^1.1.0"
+  }
 }
